fix(64-bit): 10 pointer truncations, and revert the per-arch Trace retbuf
Four issues reported by a downstream Q2PRO port of this botlib.  Two were
ours; two were not.

R-VER-3 -- pointer->int truncation, ours.  BotChat_ExitGame((int)(intptr_t)bs)
and FreeWeightConfig2((int)(intptr_t)weaponstate->weightconfig) both narrow a
pointer at a call whose callee takes a real pointer, so `sv removebot` died
inside the botlib on LP64.  The disasm settles that the original passed
pointers: BotShutdownClient pushes esi = &botstates[a1] at 0x100296C8, and
BotFreeWeaponWeights pushes eax = [esi+8] at 0x1003530C.

only sees the callee half of the class, so two source passes it is blind to
(a warning build re-enabling the four -Wno- flags BOTCFLAGS silences, plus a
grep for the double cast) swept the rest.  Eight more, three of them on the
same BotShutdownClient path:

    accessors -- on 64-bit the real lists leaked and the side-band entry
    stayed dangling for the next client in that slot;
  - BotResetState freed through the accessor but relied on memset(bs,0,..)
    to clear, which cannot reach outside bs (now cleared under
    #if BOTLIB_NEED_SIDEBAND, so the 32-bit byte match keeps its one memset);
  - BotGetMessageTeamGoal read the raw slot, so checkpoints were never found;
    ((script_t *)script)->flags on entry, making every parser diagnostic a
    crash instead of a message, across 29 call sites.  Restored script_t *,
    PC_FindDefine's return type in the same family;
  - AAS_InitRoutingUpdate's two FreeMemory calls (latent: AAS_Shutdown
    memsets aasworld, so only a second AAS_InitRouting without a shutdown
    between reaches it).

R-97 -- Trace's calling convention, ours.  ac6e9de1 gave
botimport_block_t::Trace, AAS_Trace and BotLibImport_Trace an
`#if __x86_64__ || __aarch64__` branch declaring the retbuf as an explicit
first parameter.  It worked only because both sides were ours: the published
never branched, so a foreign engine built against the published header
disagreed with our aarch64 botlib and every trace came back zeroed (AAPCS64
carries the hidden buffer in x8, not x0).  On x86-64 SysV the two spellings
are interchangeable, which is why it hid for three months.  All three
branches removed: by value on both sides, on every target -- also the
faithful 1999 spelling and the only form that reproduces gladi386.so's F663.

R-100 -- bsp_trace_t 84 vs 80 and bot_updateclient_t 1228 vs 1292, NOT ours.
Both measure the original here (84 at either word size; 0x4CC is the
`mov ecx,0x133; rep movs` in BotUpdateClient at 0x100298BE) and there is no
host side: botlib_port.h defers to a host q_shared.h's qboolean, so a C bool
shrinks the published struct to 80, and a MAX_STATS of 64 grows
bot_updateclient_t to 1292 and displaces inventory by 16 slots.  Neither is
diagnosable at runtime, so struct_sizes_asserts.h now hard-asserts
offsets, MAX_STATS==32 and sizeof(bot_updateclient_t)==0x4CC.  A port that
redefines either now fails to build with a named reason.

No fix can improve either oracle: each is on a 64-bit-only path or is a type
restoration that both 1999 compilers lower to the same dword, so the oracles
are blind to this bug class.  Zero regression is proven whole-image instead:

                        row-for-row identical to the pre-fix baseline
  gcc272 botlib .so     718 MATCH, 21520 insn_diffs -- "nothing moved"
  gcc272 game .so      1985 MATCH,   500 insn_diffs -- "nothing moved"

Every touched function still byte-exact on both oracles (AAS_Trace,
BotShutdownClient, BotFreeWeaponWeights, BotResetState, AAS_InitRoutingUpdate,
ScriptError, ScriptWarning, PS_ExpectAnyToken, ReadSignedFloat,
PC_RemoveGlobalDefine, and game-side BotLibImport_Trace).  scan_aarch64_ptr_trunc
still 0 DIRECT; check_docblocks --verify-ranges clean; native aarch64 and
i686-w64-mingw32 builds clean.  Engine smoke test (the only runtime-visible
change is Trace): survives 60 s on q2dm1, AAS initialized, bot spawns, jumps
and frags.

Memory notes updated so the reverted pattern is not reintroduced:
engine_diagnostics records that the per-arch retbuf mirror was tried and
why by-value on both sides is the fix; aarch64_pointer_truncation records the
two caller-side detector passes and the side-band accessor corollary.

Source-commit: 57ce85a3ce47e7267f3563cc4034fc4c75ed0097

+++ botlib/be_aas_route.c
@@ -351,10 +351,10 @@ void AAS_InitRoutingUpdate(void)
   /* sizeof(), not the original's 40, so each slot grows with the typed
    * struct. */
   if ( aasworld.areaupdate )
-    FreeMemory((int)(intptr_t)aasworld.areaupdate);
+    FreeMemory(aasworld.areaupdate);
   aasworld.areaupdate = (aas_routingupdate_t *)GetClearedMemory(sizeof(aas_routingupdate_t) * aasworld.numareas);
   if ( aasworld.portalupdate )
-    FreeMemory((int)(intptr_t)aasworld.portalupdate);
+    FreeMemory(aasworld.portalupdate);
   aasworld.portalupdate = (aas_routingupdate_t *)GetClearedMemory(sizeof(aas_routingupdate_t) * aasworld.numareas);
 }
